requirements.txt
fastapi
uvicorn
sqlalchemy
databases[sqlite]
python-jose[cryptography]
websockets
faker
pyjwt
passlib[bcrypt]